Added icons to AdminSideBar & other minor changes
