+++ src/app/admin/layout.tsx
@@ -1,8 +1,6 @@
 "use client";
 
 import AdminSideBar from "@/components/sidebar/adminSidebar";
-import ChatSideBar from "@/components/sidebar/chatSidebar";
-import SideBar from "@/components/sidebar/sidebar";
 import { storageAtom } from "@/store";
 import { useAtom } from "jotai";
 import { useState } from "react";
@@ -42,9 +40,7 @@ export default function Layout({ children }: { children: React.ReactNode }) {
               : "-translate-x-full absolute"
           } md:translate-x-0 md:contents items-stretch z-40 md:z-0 duration-200 ease-in-out h-screen`}
         >
-          <SideBar>
-            <AdminSideBar />
-          </SideBar>
+          <AdminSideBar />
         </div>
 
         <div
